+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Custom fasting goals now sync from phone to watch via Wearable Data Layer
+
 ### Changed
 
 ### Fixed